Seperated item store into partials
